Use bulk indexer for writes. Special case get by ID.
