+++ search/main.go
@@ -37,7 +37,7 @@ func (r *Relay) Init() error {
 func (r *Relay) AcceptEvent(evt *nostr.Event) bool {
 	// block events that are too large
 	jsonb, _ := json.Marshal(evt)
-	if len(jsonb) > 10000 {
+	if len(jsonb) > 100000 {
 		return false
 	}
